requirements.txt
pytest==8.3.5
google-cloud-resource-manager
google-cloud-firestore
functions-framework==3.*